+++ packages/ui/src/layouts/shared/files-tab/components/editor/FileEditor.vue
@@ -3,119 +3,19 @@
 		ref="editorContainer"
 		class="relative flex flex-col overflow-hidden rounded-[20px] border border-solid border-surface-4 shadow-sm"
 	>
-		<Transition name="find">
-			<div
-				v-if="isFindOpen && !isEditingImage"
-				class="absolute right-3 top-3 z-10 flex flex-col gap-1 rounded-2xl border border-solid border-surface-5 bg-surface-3 p-1.5 shadow-lg"
-				@keydown.escape.stop="closeFind"
-			>
-				<!-- Find row -->
-				<div class="flex items-center gap-1">
-					<Button
-						v-tooltip="formatMessage(messages.toggleReplace)"
-						icon-only
-						transparent
-						:aria-label="formatMessage(messages.toggleReplace)"
-						@click="toggleReplace"
-					>
-						<ChevronRightIcon
-							class="transition-transform duration-150"
-							:class="{ 'rotate-90': isReplaceOpen }"
-						/>
-					</Button>
-					<div
-						@keydown.enter.prevent.stop="findNext"
-						@keydown.shift.enter.prevent.stop="findPrevious"
-					>
-						<StyledInput
-							ref="findInputRef"
-							v-model="inFileFindQuery"
-							type="search"
-							size="small"
-							autocomplete="off"
-							:placeholder="formatMessage(messages.findInFile)"
-							wrapper-class="w-44"
-						/>
-					</div>
-					<span class="min-w-[6rem] px-1 text-right text-sm text-secondary tabular-nums">
-						{{
-							findMatchCount > 0
-								? formatMessage(messages.matchCount, {
-										current: currentFindMatch,
-										total: findMatchCount,
-									})
-								: inFileFindQuery
-									? formatMessage(messages.noResults)
-									: ''
-						}}
-					</span>
-					<Button
-						v-tooltip="formatMessage(messages.previousMatch)"
-						icon-only
-						transparent
-						:disabled="findMatchCount === 0"
-						:aria-label="formatMessage(messages.previousMatch)"
-						@click="findPrevious"
-					>
-						<ChevronUpIcon />
-					</Button>
-					<Button
-						v-tooltip="formatMessage(messages.nextMatch)"
-						icon-only
-						transparent
-						:disabled="findMatchCount === 0"
-						:aria-label="formatMessage(messages.nextMatch)"
-						@click="findNext"
-					>
-						<ChevronDownIcon />
-					</Button>
-					<div class="mx-0.5 h-4 w-px bg-surface-5" />
-					<Button
-						v-tooltip="formatMessage(messages.closeFind)"
-						icon-only
-						transparent
-						:aria-label="formatMessage(messages.closeFind)"
-						@click="closeFind"
-					>
-						<XIcon />
-					</Button>
-				</div>
-
-				<!-- Replace row -->
-				<div v-if="isReplaceOpen" class="flex items-center gap-1">
-					<div class="w-9 flex-shrink-0" />
-					<div @keydown.enter.prevent.stop="replaceOne">
-						<StyledInput
-							ref="replaceInputRef"
-							v-model="replaceQuery"
-							type="text"
-							size="small"
-							autocomplete="off"
-							:placeholder="formatMessage(messages.replaceInFile)"
-							wrapper-class="w-44"
-						/>
-					</div>
-					<ButtonStyled type="outlined">
-						<button
-							class="!h-8 whitespace-nowrap !border !border-surface-5 px-2 text-sm disabled:opacity-50"
-							:disabled="findMatchCount === 0"
-							@click="replaceOne"
-						>
-							{{ formatMessage(messages.replace) }}
-						</button>
-					</ButtonStyled>
-					<ButtonStyled type="outlined">
-						<button
-							class="!h-8 whitespace-nowrap !border !border-surface-5 px-2 text-sm disabled:opacity-50"
-							:disabled="findMatchCount === 0"
-							@click="replaceAllOccurrences"
-						>
-							{{ formatMessage(messages.replaceAll) }}
-						</button>
-					</ButtonStyled>
-				</div>
-			</div>
-		</Transition>
+		<EditorFindReplace
+			ref="findReplaceRef"
+			v-model:is-find-open="isFindOpen"
+			v-model:find-query="inFileFindQuery"
+			:is-editing-image="isEditingImage"
+			:find-match-count="findMatchCount"
+			:current-find-match="currentFindMatch"
+			@find-next="findNext"
+			@find-previous="findPrevious"
+			@close="closeFind"
+			@replace="replaceOne"
+			@replace-all="replaceAllOccurrences"
+		/>
 		<component
 			:is="props.editorComponent"
 			v-if="!isEditingImage && !isLoading && props.editorComponent"
@@ -139,24 +39,16 @@
 </template>
 
 <script setup lang="ts">
-import {
-	ChevronDownIcon,
-	ChevronRightIcon,
-	ChevronUpIcon,
-	SpinnerIcon,
-	XIcon,
-} from '@modrinth/assets'
+import { SpinnerIcon } from '@modrinth/assets'
 import { type Component, computed, nextTick, onMounted, onUnmounted, ref, watch } from 'vue'
 
-import Button from '#ui/components/base/Button.vue'
-import ButtonStyled from '#ui/components/base/ButtonStyled.vue'
-import StyledInput from '#ui/components/base/StyledInput.vue'
 import { defineMessages, useVIntl } from '#ui/composables/i18n'
 import { injectNotificationManager } from '#ui/providers/web-notifications'
 import { getEditorLanguage, getFileExtension, isImageFile } from '#ui/utils/file-extensions'
 
 import { injectFileManager } from '../../providers/file-manager'
 import type { EditingFile } from '../../types'
+import EditorFindReplace from './EditorFindReplace.vue'
 import FileImageViewer from './FileImageViewer.vue'
 
 interface MclogsResponse {
@@ -245,46 +137,6 @@ const messages = defineMessages({
 		id: 'files.editor.failed-to-share-text',
 		defaultMessage: 'Could not upload to mclo.gs.',
 	},
-	findInFile: {
-		id: 'files.editor.find-in-file',
-		defaultMessage: 'Find',
-	},
-	matchCount: {
-		id: 'files.editor.find-match-count',
-		defaultMessage: '{current} of {total}',
-	},
-	noResults: {
-		id: 'files.editor.find-no-results',
-		defaultMessage: 'No results',
-	},
-	previousMatch: {
-		id: 'files.editor.find-previous-match',
-		defaultMessage: 'Previous match',
-	},
-	nextMatch: {
-		id: 'files.editor.find-next-match',
-		defaultMessage: 'Next match',
-	},
-	closeFind: {
-		id: 'files.editor.find-close',
-		defaultMessage: 'Close',
-	},
-	toggleReplace: {
-		id: 'files.editor.find-toggle-replace',
-		defaultMessage: 'Toggle replace',
-	},
-	replaceInFile: {
-		id: 'files.editor.replace-in-file',
-		defaultMessage: 'Replace',
-	},
-	replace: {
-		id: 'files.editor.replace',
-		defaultMessage: 'Replace',
-	},
-	replaceAll: {
-		id: 'files.editor.replace-all',
-		defaultMessage: 'Replace All',
-	},
 })
 
 const fileContent = ref('')
@@ -297,13 +149,10 @@ const editorContainer = ref<HTMLElement | null>(null)
 const editorHeight = ref('300px')
 
 const isFindOpen = ref(false)
-const isReplaceOpen = ref(false)
 const inFileFindQuery = ref('')
-const replaceQuery = ref('')
 const findMatchCount = ref(0)
 const currentFindMatch = ref(0)
-const findInputRef = ref<{ focus: () => void } | null>(null)
-const replaceInputRef = ref<{ focus: () => void } | null>(null)
+const findReplaceRef = ref<{ focusFindInput: () => void; openReplace: () => void } | null>(null)
 
 watch(inFileFindQuery, handleFindInput)
 
@@ -404,8 +253,7 @@ function onEditorInit(editor: AceEditorInstance) {
 		bindKey: { win: 'Ctrl-H', mac: 'Command-Option-F' },
 		exec: () => {
 			isFindOpen.value = true
-			isReplaceOpen.value = true
-			nextTick(() => findInputRef.value?.focus())
+			nextTick(() => findReplaceRef.value?.openReplace())
 		},
 	})
 }
@@ -484,33 +332,23 @@ function toggleFind() {
 		closeFind()
 	} else {
 		isFindOpen.value = true
-		isReplaceOpen.value = false
-		nextTick(() => findInputRef.value?.focus())
-	}
-}
-
-function toggleReplace() {
-	isReplaceOpen.value = !isReplaceOpen.value
-	if (isReplaceOpen.value) {
-		nextTick(() => replaceInputRef.value?.focus())
+		nextTick(() => findReplaceRef.value?.focusFindInput())
 	}
 }
 
 function closeFind() {
 	isFindOpen.value = false
-	isReplaceOpen.value = false
 	inFileFindQuery.value = ''
-	replaceQuery.value = ''
 	findMatchCount.value = 0
 	currentFindMatch.value = 0
 	editorInstance.value?.find('', { wrap: true })
 	editorInstance.value?.focus()
 }
 
-function replaceOne() {
+function replaceOne(query: string) {
 	const editor = editorInstance.value
 	if (!editor || findMatchCount.value === 0) return
-	editor.replace(replaceQuery.value)
+	editor.replace(query)
 	nextTick(() => {
 		const count = countOccurrences(fileContent.value, inFileFindQuery.value)
 		findMatchCount.value = count
@@ -518,10 +356,10 @@ function replaceOne() {
 	})
 }
 
-function replaceAllOccurrences() {
+function replaceAllOccurrences(query: string) {
 	const editor = editorInstance.value
 	if (!editor || findMatchCount.value === 0) return
-	editor.replaceAll(replaceQuery.value)
+	editor.replaceAll(query)
 	nextTick(() => {
 		const count = countOccurrences(fileContent.value, inFileFindQuery.value)
 		findMatchCount.value = count
@@ -593,18 +431,3 @@ defineExpose({
 	toggleFind,
 })
 </script>
-
-<style scoped>
-.find-enter-active,
-.find-leave-active {
-	transition:
-		opacity 0.15s ease,
-		transform 0.15s ease;
-}
-
-.find-enter-from,
-.find-leave-to {
-	opacity: 0;
-	transform: translateY(-4px) scale(0.97);
-}
-</style>
